filter the events
also, ABI works now. It needed to be uint256 instead of uint.
According to https://github.com/ethereum/wiki/wiki/Ethereum-Contract-ABI#types
(why are there even synonyms?)

+++ imports/parityScratchpad.js
@@ -11,8 +11,8 @@ const abi = [{
   type: 'event',
   inputs: [
     { type: 'address', name: 'ofAsset', indexed: true },
-    { type: 'uint', name: 'atTimestamp' },
-    { type: 'uint', name: 'ofPrice' },
+    { type: 'uint256', name: 'atTimestamp' },
+    { type: 'uint256', name: 'ofPrice' },
   ],
   /*outputs: [
     { type: 'uint', name: 'atTimestamp' },
